refactor: 도메인 모듈 분리하면서 놓쳤던 버그 수정

+++ domain/domain-rdb/src/main/java/com/gamzabit/domain/order/OrderTransactionReader.java
@@ -16,6 +16,6 @@ public class OrderTransactionReader {
     private final OrderTransactionQueryRepository orderTransactionRepository;
 
     public List<OrderHistory> getOrderHistory(User user) {
-        return orderTransactionRepository.findAllByUserId(user.id().longValue());
+        return orderTransactionRepository.findAllByUserId(user.id());
     }
 }
